Bugfixes and more
Removerd horizontal scroll, added more labels to the overview and
adjusted some padding and widths. Made navigation using the tabs not
buggy, added actual dates and times to the comboboxes, divided the
initializermethod in the wizard into two new to make sure that we don't
do the same thing twice.